+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **[BREAKING CHANGE]** Replace `TopTripCardList` component with `TripsSection` section.
 - **[UPDATE]** Add new Pages leftnav category and a few messaging & ridedetails pages
+- **[BREAKING CHANGE]** Fix background/content classname props on `BaseSection` and `HighlightSection`
 - [...]
 
 # v19.1.1 (07/02/2020)